requirements.txt
rich
python-dotenv
openai
mcp
tiktoken
asyncio

